// UI/src/app/models/timesheet.ts
export interface Timesheet {
    _id: any;
    EmployeeId: String;
    Date: String;
    WorkDetails: String;
    Hours: String;
    Reason: String;
    Status: String;
    ApprovedBy: String;
    ApprovedReason: String;
    ApprovedDate: String;
    RejectBy: String;
    RejectDate: String;
    RejectReason: String;
}